Split calendar into company, person and department calendar table
