--- src/main/java/org/synyx/urlaubsverwaltung/calendar/Calendar.java
@@ -1,56 +1,0 @@
-package org.synyx.urlaubsverwaltung.calendar;
-
-import org.apache.commons.lang3.RandomStringUtils;
-import org.hibernate.validator.constraints.Length;
-
-import javax.persistence.Convert;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Inheritance;
-import java.time.Period;
-
-@Entity
-@Inheritance
-abstract class Calendar {
-
-    private static final int SECRET_LENGTH = 32;
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Length(min = SECRET_LENGTH, max = SECRET_LENGTH)
-    private String secret;
-
-    @Convert(converter = PeriodConverter.class)
-    private Period calendarPeriod;
-
-    protected Calendar() {
-        generateSecret();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getSecret() {
-        return secret;
-    }
-
-    public void generateSecret() {
-        secret = RandomStringUtils.randomAlphanumeric(SECRET_LENGTH);
-    }
-
-    public Period getCalendarPeriod() {
-        return calendarPeriod;
-    }
-
-    public void setCalendarPeriod(Period calendarPeriod) {
-        this.calendarPeriod = calendarPeriod;
-    }
-}
