+++ src/main/java/org/synyx/urlaubsverwaltung/calendar/DepartmentCalendar.java
@@ -1,14 +1,26 @@
 package org.synyx.urlaubsverwaltung.calendar;
 
+import org.apache.commons.lang3.RandomStringUtils;
+import org.hibernate.validator.constraints.Length;
 import org.synyx.urlaubsverwaltung.person.Person;
 
 import javax.persistence.Column;
+import javax.persistence.Convert;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 import javax.persistence.OneToOne;
 import javax.validation.constraints.NotNull;
+import java.time.Period;
 
 @Entity
-class DepartmentCalendar extends Calendar {
+class DepartmentCalendar {
+
+    private static final int SECRET_LENGTH = 32;
+
+    @Id
+    @GeneratedValue
+    private Long id;
 
     @NotNull
     @Column(name = "department_id")
@@ -18,14 +30,28 @@ class DepartmentCalendar extends Calendar {
     @OneToOne
     private Person person;
 
+    @Length(min = SECRET_LENGTH, max = SECRET_LENGTH)
+    private String secret;
+
+    @Convert(converter = PeriodConverter.class)
+    private Period calendarPeriod;
+
     public DepartmentCalendar() {
         // for hibernate - do not use this
     }
 
     DepartmentCalendar(Integer departmentId, Person person) {
-        super();
+        generateSecret();
         this.departmentId = departmentId;
         this.person = person;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public Integer getDepartmentId() {
@@ -42,5 +68,21 @@ class DepartmentCalendar extends Calendar {
 
     public void setPerson(Person person) {
         this.person = person;
+    }
+
+    public void generateSecret() {
+        secret = RandomStringUtils.randomAlphanumeric(SECRET_LENGTH);
+    }
+
+    public String getSecret() {
+        return secret;
+    }
+
+    public Period getCalendarPeriod() {
+        return calendarPeriod;
+    }
+
+    public void setCalendarPeriod(Period calendarPeriod) {
+        this.calendarPeriod = calendarPeriod;
     }
 }
