--- src/main/resources/dbchangelogs/changelog-4.14.0-add-calendar-unique-constraint.xml
@@ -1,16 +1,0 @@
-<databaseChangeLog
-  xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.10.xsd">
-
-  <changeSet author="schneider" id="add-calendar-unique-constraint">
-
-    <preConditions>
-      <tableExists tableName="calendar"/>
-    </preConditions>
-
-    <addUniqueConstraint tableName="calendar"
-                         columnNames="dtype, department_id, person_id"
-                         constraintName="unique_calendar_per_type"/>
-  </changeSet>
-</databaseChangeLog>
