+++ src/main/resources/dbchangelogs/changelogmaster.xml
@@ -76,5 +76,5 @@
   <include file="dbchangelogs/changelog-4.12.0-default-naming-strategy.xml"/>
   <include file="dbchangelogs/changelog-4.13.0-add-department-member-accession-date.xml"/>
   <include file="dbchangelogs/changelog-4.14.0-add-global-default-working-days.xml"/>
-  <include file="dbchangelogs/changelog-4.14.0-add-calendar-unique-constraint.xml"/>
+  <include file="dbchangelogs/changelog-4.14.0-split-calendar-tables.xml"/>
 </databaseChangeLog>
